requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

aiofiles==24.1.0
aiogram==3.20.0.post0
annotated-types==0.7.0
certifi==2025.4.26
jupyter_core==5.8.1
magic-filter==1.0.12
pigar==2.1.7
pip==24.0
platformdirs==4.3.8
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
typing-inspection==0.4.0
